+++ test/unit/aws_encryption_sdk/internal/formatting/test_deserialize.py
@@ -23,7 +23,7 @@ from aws_encryption_sdk.exceptions import NotSupportedError, SerializationError,
 from aws_encryption_sdk.identifiers import AlgorithmSuite
 from aws_encryption_sdk.internal.structures import EncryptedData
 
-from .test_values import VALUES
+from ....test_values import VALUES
 
 pytestmark = [pytest.mark.unit, pytest.mark.local]
