+++ test/unit/aws_encryption_sdk/internal/formatting/test_encryption_context.py
@@ -18,7 +18,7 @@ import aws_encryption_sdk.internal.formatting.encryption_context
 from aws_encryption_sdk.exceptions import SerializationError
 from aws_encryption_sdk.identifiers import ContentAADString
 
-from .test_values import VALUES
+from ....test_values import VALUES
 
 pytestmark = [pytest.mark.unit, pytest.mark.local]
