+++ test/unit/aws_encryption_sdk/internal/test_defaults.py
