+++ test/unit/aws_encryption_sdk/internal/test_structures.py
@@ -21,7 +21,7 @@ from aws_encryption_sdk.internal.structures import (
     MessageNoFrameBody,
 )
 
-from .unit_test_utils import all_invalid_kwargs, all_valid_kwargs
+from ...unit_test_utils import all_invalid_kwargs, all_valid_kwargs
 
 pytestmark = [pytest.mark.unit, pytest.mark.local]
