+++ test/unit/aws_encryption_sdk/key_providers/base/test_base_master_key_config.py
@@ -15,7 +15,7 @@ import pytest
 
 from aws_encryption_sdk.key_providers.base import MasterKeyConfig
 
-from .unit_test_utils import all_invalid_kwargs, all_valid_kwargs
+from ....unit_test_utils import all_invalid_kwargs, all_valid_kwargs
 
 pytestmark = [pytest.mark.unit, pytest.mark.local]
