+++ test/unit/aws_encryption_sdk/key_providers/kms/test_kms_master_key_config.py
@@ -17,7 +17,7 @@ import pytest
 from aws_encryption_sdk.key_providers.base import MasterKeyConfig
 from aws_encryption_sdk.key_providers.kms import _PROVIDER_ID, KMSMasterKeyConfig
 
-from .unit_test_utils import all_invalid_kwargs, all_valid_kwargs
+from ....unit_test_utils import all_invalid_kwargs, all_valid_kwargs
 
 pytestmark = [pytest.mark.unit, pytest.mark.local]
