+++ test/unit/aws_encryption_sdk/key_providers/kms/test_kms_master_key_provider_config.py
@@ -17,7 +17,7 @@ import pytest
 from aws_encryption_sdk.key_providers.base import MasterKeyProviderConfig
 from aws_encryption_sdk.key_providers.kms import KMSMasterKeyProviderConfig
 
-from .unit_test_utils import all_invalid_kwargs, all_valid_kwargs
+from ....unit_test_utils import all_invalid_kwargs, all_valid_kwargs
 
 pytestmark = [pytest.mark.unit, pytest.mark.local]
