+++ test/unit/aws_encryption_sdk/materials_managers/test_default.py
