+++ test/unit/aws_encryption_sdk/streaming_client/test_configs.py
@@ -22,7 +22,7 @@ from aws_encryption_sdk.materials_managers.base import CryptoMaterialsManager
 from aws_encryption_sdk.materials_managers.default import DefaultCryptoMaterialsManager
 from aws_encryption_sdk.streaming_client import DecryptorConfig, EncryptorConfig, _ClientConfig
 
-from .unit_test_utils import all_invalid_kwargs, all_valid_kwargs, build_valid_kwargs_list
+from ...unit_test_utils import all_invalid_kwargs, all_valid_kwargs, build_valid_kwargs_list
 
 pytestmark = [pytest.mark.unit, pytest.mark.local]
