+++ test/unit/aws_encryption_sdk/streaming_client/test_stream_encryptor.py
@@ -30,7 +30,7 @@ from aws_encryption_sdk.materials_managers.base import CryptoMaterialsManager
 from aws_encryption_sdk.streaming_client import StreamEncryptor
 from aws_encryption_sdk.structures import MessageHeader
 
-from .test_values import VALUES
+from ...test_values import VALUES
 
 pytestmark = [pytest.mark.unit, pytest.mark.local]
