+++ test/unit/aws_encryption_sdk/test_client.py
